chore: rename module and change import statements
