+++ powersimdata/input/tests/test_grid.py
@@ -7,7 +7,7 @@ import pytest
 from powersimdata.input.grid import Grid
 from powersimdata.input.helpers import add_column_to_data_frame
 from powersimdata.input.scenario_grid import format_gencost, link
-from powersimdata.network.usa_tamu.usa_tamu_model import TAMU, check_interconnect
+from powersimdata.network.usa_tamu.model import TAMU, check_interconnect
 
 
 def test_interconnect_type():
